Versão final

+++ Core/Inc/Servo.h
@@ -9,16 +9,13 @@
 #define INC_SERVO_H_
 
 
-//#define PERIOD 1250
-//#define CHANNEL TIM_CHANNEL_2
-
 #include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include "stm32f1xx_hal.h"
 
 
-void setServoAngle(TIM_HandleTypeDef* timer, uint32_t channel, uint16_t period, int16_t angle);
+void setServoAngle(TIM_HandleTypeDef* timer, uint32_t channel, uint16_t period, int8_t angle);
 void setServoPWM(TIM_HandleTypeDef* timer, uint32_t channel, uint16_t period, uint16_t pulse);
 
 #endif /* INC_SERVO_H_ */
